update ollama & langchain, fix payload test
