+++ Cargo.toml
@@ -42,8 +42,8 @@ sha2 = "0.10.8"
 sha3 = "0.10.8"
 
 # llm stuff
-langchain-rust = { version = "4.2.0", features = ["ollama"] }
-ollama-rs = "0.1.9"
+langchain-rust = { version = "4.3.0", features = ["ollama"] }
+ollama-rs = "0.2.0"
 ollama-workflows = { git = "https://github.com/andthattoo/ollama-workflows", branch = "main" }
 
 uuid = { version = "1.8.0", features = ["v4"] }
@@ -51,9 +51,6 @@ uuid = { version = "1.8.0", features = ["v4"] }
 [dev-dependencies]
 colored = "2.1.0"
 rand = "0.8.5"
-
-# TODO: fixed version of ollama-rs for benchmarks, remove this when the new version is released
-ollama-rs-master = { package = "ollama-rs", git = "https://github.com/pepperoni21/ollama-rs.git", branch = "master" }
 
 [[example]]
 name = "ollama"
