+++ examples/common/ollama.rs
@@ -2,7 +2,7 @@ use dkn_compute::{
     config::constants::{DEFAULT_OLLAMA_HOST, DEFAULT_OLLAMA_PORT},
     utils::get_current_time_nanos,
 };
-use ollama_rs_master::{
+use ollama_rs::{
     generation::completion::{request::GenerationRequest, GenerationResponse},
     Ollama,
 };
@@ -12,7 +12,7 @@ pub async fn use_model_with_prompt(
     model: &str,
     prompt: &str,
 ) -> (GenerationResponse, tokio::time::Duration) {
-    let ollama = Ollama::new(DEFAULT_OLLAMA_HOST, DEFAULT_OLLAMA_PORT);
+    let ollama = Ollama::new(DEFAULT_OLLAMA_HOST.to_string(), DEFAULT_OLLAMA_PORT);
 
     let time = get_current_time_nanos();
     let prompt = prompt.to_string();
