+++ src/config/mod.rs
@@ -71,11 +71,9 @@ impl DriaComputeNodeConfig {
             serde_json::to_string(&models).unwrap_or_default()
         );
 
-        if !cfg!(test) {
-            assert!(
-                !models.is_empty(),
-                "At least one model should be provided in the configuration."
-            );
+        if models.is_empty() {
+            log::error!("No models were provided, you will not be able to get any tasks!");
+            log::error!("Please restart with at least one model provided within DKN_MODELS.");
         }
 
         Self {
