+++ tests/compute_test.rs
@@ -1,45 +1,44 @@
 #![allow(unused_imports)]
 
-mod compute_test {
-    use langchain_rust::{language_models::llm::LLM, llm::client::Ollama};
-    use ollama_workflows::{Entry, Executor, Model, ProgramMemory, Workflow};
-    use std::env;
-    use tokio_util::sync::CancellationToken;
+use langchain_rust::{language_models::llm::LLM, llm::client::Ollama};
+use ollama_workflows::{Entry, Executor, Model, ProgramMemory, Workflow};
+use std::env;
+use tokio_util::sync::CancellationToken;
 
-    #[tokio::test]
-    #[ignore = "run this manually"]
-    async fn test_ollama_prompt() {
-        let model = "orca-mini".to_string();
-        let ollama = Ollama::default().with_model(model);
-        let prompt = "The sky appears blue during the day because of a process called scattering. \
+#[tokio::test]
+#[ignore = "run this manually"]
+async fn test_ollama_prompt() {
+    let model = "orca-mini".to_string();
+    let ollama = Ollama::default().with_model(model);
+    let prompt = "The sky appears blue during the day because of a process called scattering. \
                 When sunlight enters the Earth's atmosphere, it collides with air molecules such as oxygen and nitrogen. \
                 These collisions cause some of the light to be absorbed or reflected, which makes the colors we see appear more vivid and vibrant. \
                 Blue is one of the brightest colors that is scattered the most by the atmosphere, making it visible to our eyes during the day. \
                 What may be the question this answer?".to_string();
 
-        let response = ollama
-            .invoke(&prompt)
-            .await
-            .expect("Should generate response");
-        println!("Prompt: {}\n\nResponse:{}", prompt, response);
-    }
+    let response = ollama
+        .invoke(&prompt)
+        .await
+        .expect("Should generate response");
+    println!("Prompt: {}\n\nResponse:{}", prompt, response);
+}
 
-    #[tokio::test]
-    #[ignore = "run this manually"]
-    async fn test_ollama_bad_model() {
-        let model = "thismodeldoesnotexistlol".to_string();
-        let ollama = Ollama::default().with_model(model);
-        let setup_res = ollama.invoke("hola").await;
-        assert!(
-            setup_res.is_err(),
-            "Should give error due to non-existing model."
-        );
-    }
+#[tokio::test]
+#[ignore = "run this manually"]
+async fn test_ollama_bad_model() {
+    let model = "thismodeldoesnotexistlol".to_string();
+    let ollama = Ollama::default().with_model(model);
+    let setup_res = ollama.invoke("hola").await;
+    assert!(
+        setup_res.is_err(),
+        "Should give error due to non-existing model."
+    );
+}
 
-    #[tokio::test]
-    #[ignore = "run this manually"]
-    async fn test_workflow() {
-        let workflow = r#"{
+#[tokio::test]
+#[ignore = "run this manually"]
+async fn test_workflow() {
+    let workflow = r#"{
     "name": "Simple",
     "description": "This is a simple workflow",
     "config": {
@@ -76,13 +75,12 @@ mod compute_test {
         }
     ]
 }"#;
-        let workflow: Workflow = serde_json::from_str(workflow).unwrap();
-        let exe = Executor::new(Model::Phi3Mini);
-        let mut memory = ProgramMemory::new();
+    let workflow: Workflow = serde_json::from_str(workflow).unwrap();
+    let exe = Executor::new(Model::Phi3Mini);
+    let mut memory = ProgramMemory::new();
 
-        exe.execute(None, workflow, &mut memory).await;
+    exe.execute(None, workflow, &mut memory).await;
 
-        let result = memory.read(&"final_result".to_string()).unwrap();
-        println!("Result: {}", result);
-    }
+    let result = memory.read(&"final_result".to_string()).unwrap();
+    println!("Result: {}", result);
 }
